fix for code review
