+++ model/patch_lifecycle.go
@@ -2,6 +2,7 @@ package model
 
 import (
 	"context"
+	"encoding/base64"
 	"fmt"
 	"io"
 	"io/ioutil"
@@ -102,6 +103,75 @@ func AddNewTasksForPatch(p *patch.Patch, patchVersion *Version, project *Project
 func IncludePatchDependencies(project *Project, tvpairs []TVPair) []TVPair {
 	di := &dependencyIncluder{Project: project}
 	return di.Include(tvpairs)
+}
+
+// GetPatchedProject creates and validates a project created by fetching latest commit information from GitHub
+// and applying the patch to the latest remote configuration. The error returned can be a validation error.
+func GetPatchedProject(ctx context.Context, p *patch.Patch, githubOauthToken string) (*Project, error) {
+	if p.Version != "" {
+		return nil, errors.Errorf("Patch %v already finalized", p.Version)
+	}
+
+	projectRef, err := FindOneProjectRef(p.Project)
+	if err != nil {
+		return nil, errors.WithStack(err)
+	}
+
+	project := &Project{}
+	// if the patched config exists, use that as the project file bytes.
+	if p.PatchedConfig != "" {
+		if err := LoadProjectInto([]byte(p.PatchedConfig), projectRef.Identifier, project); err != nil {
+			return nil, errors.WithStack(err)
+		}
+		return project, nil
+	}
+
+	var cancel context.CancelFunc
+	ctx, cancel = context.WithTimeout(ctx, 10*time.Second)
+	defer cancel()
+
+	// try to get the remote project file data at the requested revision
+	var projectFileBytes []byte
+	hash := p.Githash
+	if p.IsGithubPRPatch() {
+		hash = p.GithubPatchData.HeadHash
+	}
+	if p.IsPRMergePatch() {
+		hash = p.GithubPatchData.MergeCommitSHA
+	}
+
+	githubFile, err := thirdparty.GetGithubFile(ctx, githubOauthToken, projectRef.Owner,
+		projectRef.Repo, projectRef.RemotePath, hash)
+	if err != nil {
+		// if the project file doesn't exist, but our patch includes a project file,
+		// we try to apply the diff and proceed.
+		if !(p.ConfigChanged(projectRef.RemotePath) && thirdparty.IsFileNotFound(err)) {
+			// return an error if the github error is network/auth-related or we aren't patching the config
+			return nil, errors.Wrapf(err, "Could not get github file at '%s/%s'@%s: %s", projectRef.Owner,
+				projectRef.Repo, projectRef.RemotePath, hash)
+		}
+	} else {
+		// we successfully got the project file in base64, so we decode it
+		projectFileBytes, err = base64.StdEncoding.DecodeString(*githubFile.Content)
+		if err != nil {
+			return nil, errors.Wrapf(err, "Could not decode github file at '%s/%s'@%s: %s", projectRef.Owner,
+				projectRef.Repo, projectRef.RemotePath, hash)
+		}
+	}
+
+	// apply remote configuration patch if needed
+	if !(p.IsGithubPRPatch() || p.IsPRMergePatch()) && p.ConfigChanged(projectRef.RemotePath) {
+		projectFileBytes, err = MakePatchedConfig(ctx, p, projectRef.RemotePath, string(projectFileBytes))
+		if err != nil {
+			return nil, errors.Wrapf(err, "Could not patch remote configuration file")
+		}
+	}
+
+	if err := LoadProjectInto(projectFileBytes, projectRef.Identifier, project); err != nil {
+		return nil, errors.WithStack(err)
+	}
+
+	return project, nil
 }
 
 // MakePatchedConfig takes in the path to a remote configuration a stringified version
