+++ model/patch_lifecycle_test.go
@@ -6,19 +6,272 @@ import (
 	"io/ioutil"
 	"os"
 	"path/filepath"
+	"strings"
 	"testing"
 	"time"
 
 	"github.com/evergreen-ci/evergreen"
 	"github.com/evergreen-ci/evergreen/db"
 	"github.com/evergreen-ci/evergreen/model/build"
+	"github.com/evergreen-ci/evergreen/model/distro"
 	"github.com/evergreen-ci/evergreen/model/patch"
 	"github.com/evergreen-ci/evergreen/model/task"
 	"github.com/evergreen-ci/evergreen/testutil"
 	"github.com/mongodb/grip"
 	. "github.com/smartystreets/goconvey/convey"
 	"github.com/stretchr/testify/assert"
+	"gopkg.in/mgo.v2/bson"
+	yaml "gopkg.in/yaml.v2"
 )
+
+var (
+	patchTestConfig = testutil.TestConfig()
+	configFilePath  = "testing/mci.yml"
+	patchedProject  = "mci-config"
+	patchedRevision = "582257a4ca3a9c890959b04d4dd2de5e4d34e9e7"
+	patchFile       = "testdata/patch2.diff"
+	patchOwner      = "deafgoat"
+	patchRepo       = "config"
+	patchBranch     = "master"
+
+	// newProjectPatchFile is a diff that adds a new project configuration file
+	// located at newConfigFilePath.
+	newProjectPatchFile = "testdata/project2.diff"
+	newConfigFilePath   = "testing/project2.config"
+)
+
+func init() {
+	current := testutil.GetDirectoryOfFile()
+	patchFile = filepath.Join(current, patchFile)
+	newProjectPatchFile = filepath.Join(current, newProjectPatchFile)
+}
+
+func clearAll(t *testing.T) {
+	testutil.HandleTestingErr(
+		db.ClearCollections(
+			ProjectRefCollection,
+			patch.Collection,
+			VersionCollection,
+			build.Collection,
+			task.Collection,
+			distro.Collection,
+		), t, "Error clearing test collection")
+}
+
+// resetPatchSetup clears the ProjectRef, Patch, Version, Build, and Task Collections
+// and creates a patch from the test path given.
+func resetPatchSetup(t *testing.T, testPath string) *patch.Patch {
+	clearAll(t)
+	projectRef := &ProjectRef{
+		Identifier: patchedProject,
+		RemotePath: configFilePath,
+		Owner:      patchOwner,
+		Repo:       patchRepo,
+		Branch:     patchBranch,
+	}
+	// insert distros to be used
+	distros := []distro.Distro{{Id: "d1"}, {Id: "d2"}}
+	for _, d := range distros {
+		err := d.Insert()
+		testutil.HandleTestingErr(err, t, "Couldn't insert test distro: %v", err)
+	}
+
+	err := projectRef.Insert()
+	testutil.HandleTestingErr(err, t, "Couldn't insert test project ref: %v", err)
+
+	baseVersion := &Version{
+		Identifier: patchedProject,
+		CreateTime: time.Now(),
+		Revision:   patchedRevision,
+		Requester:  evergreen.RepotrackerVersionRequester,
+	}
+	err = baseVersion.Insert()
+	testutil.HandleTestingErr(err, t, "Couldn't insert test base version: %v", err)
+
+	fileBytes, err := ioutil.ReadFile(patchFile)
+	testutil.HandleTestingErr(err, t, "Couldn't read patch file: %v", err)
+
+	// this patch adds a new task to the existing build
+	configPatch := &patch.Patch{
+		Id:            "52549c143122",
+		Project:       patchedProject,
+		Githash:       patchedRevision,
+		Tasks:         []string{"taskTwo", "taskOne"},
+		BuildVariants: []string{"linux-64-duroff"},
+		Patches: []patch.ModulePatch{
+			{
+				Githash: "revision",
+				PatchSet: patch.PatchSet{
+					Patch: fmt.Sprintf(string(fileBytes), testPath, testPath, testPath, testPath),
+					Summary: []patch.Summary{
+						{Name: configFilePath, Additions: 4, Deletions: 80},
+						{Name: "random.txt", Additions: 6, Deletions: 0},
+					},
+				},
+			},
+		},
+	}
+	err = configPatch.Insert()
+	testutil.HandleTestingErr(err, t, "Couldn't insert test patch: %v", err)
+	return configPatch
+}
+
+func resetProjectlessPatchSetup(t *testing.T) *patch.Patch {
+	clearAll(t)
+	projectRef := &ProjectRef{
+		Identifier: patchedProject,
+		RemotePath: newConfigFilePath,
+		Owner:      patchOwner,
+		Repo:       patchRepo,
+		Branch:     patchBranch,
+	}
+	// insert distros to be used
+	distros := []distro.Distro{{Id: "d1"}, {Id: "d2"}}
+	for _, d := range distros {
+		err := d.Insert()
+		testutil.HandleTestingErr(err, t, "Couldn't insert test distro: %v", err)
+	}
+
+	err := projectRef.Insert()
+	testutil.HandleTestingErr(err, t, "Couldn't insert test project ref: %v", err)
+
+	fileBytes, err := ioutil.ReadFile(newProjectPatchFile)
+	testutil.HandleTestingErr(err, t, "Couldn't read patch file: %v", err)
+
+	// this patch adds a new task to the existing build
+	configPatch := &patch.Patch{
+		Id:            "52549c143123",
+		Project:       patchedProject,
+		BuildVariants: []string{"linux-64-duroff"},
+		Githash:       patchedRevision,
+		Patches: []patch.ModulePatch{
+			{
+				Githash: "revision",
+				PatchSet: patch.PatchSet{
+					Patch:   string(fileBytes),
+					Summary: []patch.Summary{{Name: newConfigFilePath}},
+				},
+			},
+		},
+	}
+	err = configPatch.Insert()
+	testutil.HandleTestingErr(err, t, "Couldn't insert test patch: %v", err)
+	return configPatch
+}
+
+func TestGetPatchedProject(t *testing.T) {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	testutil.ConfigureIntegrationTest(t, patchTestConfig, "TestConfigurePatch")
+	Convey("With calling GetPatchedProject with a config and remote configuration path",
+		t, func() {
+			Convey("Calling GetPatchedProject returns a valid project given a patch and settings", func() {
+				configPatch := resetPatchSetup(t, configFilePath)
+				token, err := patchTestConfig.GetGithubOauthToken()
+				So(err, ShouldBeNil)
+				project, err := GetPatchedProject(ctx, configPatch, token)
+				So(err, ShouldBeNil)
+				So(project, ShouldNotBeNil)
+			})
+
+			Convey("Calling GetPatchedProject on a project-less version returns a valid project", func() {
+				configPatch := resetProjectlessPatchSetup(t)
+				token, err := patchTestConfig.GetGithubOauthToken()
+				So(err, ShouldBeNil)
+				project, err := GetPatchedProject(ctx, configPatch, token)
+				So(err, ShouldBeNil)
+				So(project, ShouldNotBeNil)
+			})
+
+			Convey("Calling GetPatchedProject on a patch with GridFS patches works", func() {
+				configPatch := resetProjectlessPatchSetup(t)
+
+				patchFileID := bson.NewObjectId()
+				So(db.WriteGridFile(patch.GridFSPrefix, patchFileID.Hex(), strings.NewReader(configPatch.Patches[0].PatchSet.Patch)), ShouldBeNil)
+				configPatch.Patches[0].PatchSet.Patch = ""
+				configPatch.Patches[0].PatchSet.PatchFileId = patchFileID.Hex()
+
+				token, err := patchTestConfig.GetGithubOauthToken()
+				So(err, ShouldBeNil)
+				project, err := GetPatchedProject(ctx, configPatch, token)
+				So(err, ShouldBeNil)
+				So(project, ShouldNotBeNil)
+			})
+
+			Reset(func() {
+				So(db.Clear(distro.Collection), ShouldBeNil)
+				So(db.ClearGridCollections(patch.GridFSPrefix), ShouldBeNil)
+			})
+		})
+}
+
+func TestFinalizePatch(t *testing.T) {
+	testutil.ConfigureIntegrationTest(t, patchTestConfig, "TestFinalizePatch")
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	Convey("With FinalizePatch on a project and commit event generated from GetPatchedProject path",
+		t, func() {
+			configPatch := resetPatchSetup(t, configFilePath)
+			Convey("a patched config should drive version creation", func() {
+				token, err := patchTestConfig.GetGithubOauthToken()
+				So(err, ShouldBeNil)
+				project, err := GetPatchedProject(ctx, configPatch, token)
+				So(err, ShouldBeNil)
+				yamlBytes, err := yaml.Marshal(project)
+				So(err, ShouldBeNil)
+				configPatch.PatchedConfig = string(yamlBytes)
+				token, err = patchTestConfig.GetGithubOauthToken()
+				So(err, ShouldBeNil)
+				version, err := FinalizePatch(ctx, configPatch, evergreen.PatchVersionRequester, token)
+				So(err, ShouldBeNil)
+				So(version, ShouldNotBeNil)
+				// ensure the relevant builds/tasks were created
+				builds, err := build.Find(build.All)
+				So(err, ShouldBeNil)
+				So(len(builds), ShouldEqual, 1)
+				So(len(builds[0].Tasks), ShouldEqual, 2)
+				tasks, err := task.Find(task.All)
+				So(err, ShouldBeNil)
+				So(len(tasks), ShouldEqual, 2)
+			})
+
+			Convey("a patch that does not include the remote config should not "+
+				"drive version creation", func() {
+				patchedConfigFile := "fakeInPatchSoNotPatched"
+				configPatch := resetPatchSetup(t, patchedConfigFile)
+				token, err := patchTestConfig.GetGithubOauthToken()
+				So(err, ShouldBeNil)
+				project, err := GetPatchedProject(ctx, configPatch, token)
+				So(err, ShouldBeNil)
+				yamlBytes, err := yaml.Marshal(project)
+				So(err, ShouldBeNil)
+				configPatch.PatchedConfig = string(yamlBytes)
+				token, err = patchTestConfig.GetGithubOauthToken()
+				So(err, ShouldBeNil)
+				version, err := FinalizePatch(ctx, configPatch, evergreen.PatchVersionRequester, token)
+				So(err, ShouldBeNil)
+				So(version, ShouldNotBeNil)
+				So(err, ShouldBeNil)
+				So(version, ShouldNotBeNil)
+
+				// ensure the relevant builds/tasks were created
+				builds, err := build.Find(build.All)
+				So(err, ShouldBeNil)
+				So(len(builds), ShouldEqual, 1)
+				So(len(builds[0].Tasks), ShouldEqual, 1)
+				tasks, err := task.Find(task.All)
+				So(err, ShouldBeNil)
+				So(len(tasks), ShouldEqual, 1)
+			})
+
+			Reset(func() {
+				So(db.Clear(distro.Collection), ShouldBeNil)
+			})
+		})
+}
 
 func TestMakePatchedConfig(t *testing.T) {
 	ctx, cancel := context.WithCancel(context.Background())
