+++ repotracker/repotracker.go
@@ -590,10 +590,7 @@ func CreateVersionFromConfig(ref *model.ProjectRef, config *model.Project, metad
 	v.Ignored = ignore
 
 	// validate the project
-	verrs, err := validator.CheckProjectSyntax(config)
-	if err != nil {
-		return nil, errors.Wrap(err, "error validating project")
-	}
+	verrs := validator.CheckProjectSyntax(config)
 	if len(verrs) > 0 || versionErrs != nil {
 		// We have syntax errors in the project.
 		// Format them, as we need to store + display them to the user
